Fix int32 overflow in paged MQA logits block addressing
The unified KV pool hands per-layer cache views a large block stride
(~1e6 elements on DeepSeek-V4-Flash TP=8). `block_idx` loads as int32,
so `block_idx * stride_kvf_block` wraps past ~2k pool blocks, producing
a negative offset: an illegal memory access when the wrapped address is
unmapped, silent logit corruption when it is not. Cast the block index
to int64 before the multiply, matching every other pool-accessing
Triton kernel in the tree, and add a strided-pool parity regression
test that fails without the cast.

+++ vllm/v1/attention/ops/mqa_logits_triton.py
@@ -102,9 +102,12 @@ def _fp8_paged_mqa_logits_kernel(
 
     q_offset = context_len - next_n + next_n_id
 
+    # int64: unified-KV-pool layer views carry a large block stride (~1e6
+    # elements), so int32 `block_idx * stride` wraps once a batch touches
+    # roughly >2k pool blocks (long context or concurrent sequences).
     block_idx = tl.load(
         block_tables_ptr + batch_id * stride_bt_b + block_rk * stride_bt_k
-    )
+    ).to(tl.int64)
 
     offs_h = tl.arange(0, BLOCK_H)
     offs_d = tl.arange(0, BLOCK_D)
